Remove dx-run-app-locally
