+++ .gitignore
@@ -57,7 +57,6 @@ dxR.Rcheck
 /bin/dx-mappings-to-sam
 /bin/dx-mount
 /bin/dx-reads-to-fastq
-/bin/dx-run-app-locally
 /bin/dx-spans-to-bed
 /bin/dx-variants-to-vcf
 /bin/dx-verify-file
