--- doc/examples/dx-apps/cross_platform_debugging_example/.gitignore
@@ -1,5 +1,0 @@
-resources/usr/bin/
-resources/usr/lib/
-mac_resources/usr/bin/
-mac_resources/usr/lib/
-src/greeter/*.o
