--- doc/examples/dx-apps/cross_platform_debugging_example/Readme.developer.md
@@ -1,23 +1,0 @@
-# cross_platform_debugging_example Developer Readme
-
-## Running this app with additional computational resources
-
-This app has the following entry points:
-
-* main
-
-When running this app, you can override the instance type to be used by
-providing the ``systemRequirements`` field to ```/applet-XXXX/run``` or
-```/app-XXXX/run```, as follows:
-
-    {
-      systemRequirements: {
-        "main": {"instanceType": "mem2_hdd2_x2"}
-      },
-      [...]
-    }
-
-See <a
-href="https://wiki.dnanexus.com/API-Specification-v1.0.0/IO-and-Run-Specifications#Run-Specification">Run
-Specification</a> in the API documentation for more information about the
-available instance types.
