--- doc/examples/dx-apps/cross_platform_debugging_example/resources/opt/mydata
@@ -1,1 +1,0 @@
-**Successfully found data file**
