--- doc/examples/dx-apps/cross_platform_debugging_example/src/cross_platform_debugging_example.sh
@@ -1,12 +1,0 @@
-#!/bin/bash
-# cross_platform_debugging_example 0.0.1
-# Generated by dx-app-wizard.
-
-main() {
-    # Demonstrate that we can find a data file that lives in a fixed directory.
-    cat "${DX_FS_ROOT}/opt/mydata"
-
-    # Demonstrate that we can find an executable that lives in
-    # resources/usr/bin (or possibly in mac_resources/usr/bin).
-    dx-jobutil-add-output greeting "`greeter $name`" --class=string
-}
