--- doc/examples/dx-apps/cross_platform_debugging_example/src/greeter/greeter.cpp
@@ -1,7 +1,0 @@
-#include <iostream>
-
-using namespace std;
-
-void greet(const char* name) {
-  cout << "Hello, " << name << "!" << endl;
-}
