--- doc/examples/dx-apps/cross_platform_debugging_example/src/greeter/greeter.hpp
@@ -1,7 +1,0 @@
-#ifndef GREETER_HPP
-#define GREETER_HPP
-
-// Prints a greeting.
-void greet(const char*);
-
-#endif
