--- doc/examples/dx-apps/cross_platform_debugging_example/src/greeter/main.cpp
@@ -1,7 +1,0 @@
-#include "greeter.hpp"
-
-int main(int argc, char* argv[]) {
-  if (argc >= 2) {
-    greet(argv[1]);
-  }
-}
